Greetings_of_the_day after checking valid inputs
This C++ code takes an input of a Date and time in DDMMYYYY & HHMM formats respectively. Post validation (multiple: correct number of days in a month, the correct number of months in a year, and leap year checks among others) generated a predetermined output message (based upon time of the day).

The challenge was to code this without a for loop.